Added detect.cpp code to be tested
The code subscribes to scanner data and processes it. The node will
determine the urgency of an object's distance away from the copter. The
Urgency value will then be published on vel_cmd/FlightCmd.

Future implementation:
Subscriber will be added to avoid.cpp to subscribe to the urgency data.
Depending on the urgency value, velocity commands will be sent to the servos
to steer the copters forward speed and direction.